[NO-ISSUE] error 정의
