--- Projects/Core/RestAPIErrorCore/Interface/Interface.swift
@@ -1,1 +1,0 @@
-// Interface module source code
