+++ Projects/Core/UserAPICore/Project.swift
@@ -14,6 +14,7 @@ let project = Project.module(
 			dependencies: [
 				.core(target: .UserAPICore, type: .interface),
                 .core(target: .StatusCodeCore, type: .interface),
+                .core(target: .RestAPIErrorCore, type: .interface),
                 .SPM.alamofire,
                 .SPM.moya,
                 .SPM.swinject,
